feat: api/user/withdraw api 구현 - session 에 user 정보 존재 확인 - session 에 있는 user 정보가 DB 정보와 일치하는지 확인 - 불일치 -> 에러 반환 - 일치 -> 유저정보 삭제

+++ server/src/user/user.controller.ts
@@ -72,14 +72,14 @@ export class UserController {
   // 회원 탈퇴
   @Delete('/withdraw')
   withdraw(@Req() req: Request, @Res() res: Response) {
+    if (!req.session.user) {
+      throw errors.NOT_LOGGED_IN;
+    }
+
     const userInfo: UserInfo = req.session.user;
 
     this.userService.remove(userInfo);
 
-    // 응답
-    return res.status(200).send({
-      code: 10000,
-      message: '성공',
-    });
+    return res.status(200).send(new SuccessResponse());
   }
 }
